improve: transform some parameters of String to Int

+++ app/controllers/Application.scala
@@ -20,16 +20,16 @@ object Application extends Controller {
 
             .withDate( parser.retrieveValueFromPattern( LogPattern.DATE ).getOrElse( "none" ) )
             .withDuration( parser.retrieveValueFromPattern( LogPattern.BATCH_DURATION ).getOrElse( "none" ) )
-            .withRetrieve( parser.retrieveValueFromPattern( LogPattern.RETRIEVE_EVENTS ).getOrElse( "none" ) )
-            .withCallback( parser.retrieveValueFromPattern( LogPattern.CALLBACK_EVENTS ).getOrElse( "none" ) )
+            .withRetrieve( parser.retrieveValueFromPattern( LogPattern.RETRIEVE_EVENTS ).getOrElse( "none" ).toInt )
+            .withCallback( parser.retrieveValueFromPattern( LogPattern.CALLBACK_EVENTS ).getOrElse( "none" ).toInt )
 
-            .withAccountUploaded( parser.retrieveValueFromPattern( LogPattern.ACCOUNT_UPLOAD ).getOrElse( "none" ) )
-            .withContactUploaded( parser.retrieveValueFromPattern( LogPattern.CONTACT_UPLOAD ).getOrElse( "none" ) )
-            .withOpportunityUploaded( parser.retrieveValueFromPattern( LogPattern.OPPORTUNITY_UPLOAD ).getOrElse( "none" ) )
+            .withAccountUploaded( parser.retrieveValueFromPattern( LogPattern.ACCOUNT_UPLOAD ).getOrElse( "none" ).toInt )
+            .withContactUploaded( parser.retrieveValueFromPattern( LogPattern.CONTACT_UPLOAD ).getOrElse( "none" ).toInt )
+            .withOpportunityUploaded( parser.retrieveValueFromPattern( LogPattern.OPPORTUNITY_UPLOAD ).getOrElse( "none" ).toInt )
 
-            .withAccountDownloaded( parser.retrieveValueFromPattern( LogPattern.ACCOUNT_DOWNLOAD ).getOrElse( "none" ) )
-            .withContactDownloaded( parser.retrieveValueFromPattern( LogPattern.CONTACT_DOWNLOAD ).getOrElse( "none" ) )
-            .withOpportunityDownloaded( parser.retrieveValueFromPattern( LogPattern.OPPORTUNITY_DOWNLOAD ).getOrElse( "none" ) )
+            .withAccountDownloaded( parser.retrieveValueFromPattern( LogPattern.ACCOUNT_DOWNLOAD ).getOrElse( "none" ).toInt )
+            .withContactDownloaded( parser.retrieveValueFromPattern( LogPattern.CONTACT_DOWNLOAD ).getOrElse( "none" ).toInt )
+            .withOpportunityDownloaded( parser.retrieveValueFromPattern( LogPattern.OPPORTUNITY_DOWNLOAD ).getOrElse( "none" ).toInt )
             .toLog
 
         Ok(views.html.loaderEvents("batch loader du 25 mai 2012", log))
